todoist_notify/requirements.txt
requests
todoist-python
